Datos

+++ database/migrations/2022_10_15_024809_create_proveedors_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->engine = 'InnoDB';
             $table->id();
             $table->string("FullName");
-            $table->integer("Ruc");
+            $table->BigInteger("Ruc");
             $table->timestamps();
         });
     }
