+++ database/migrations/2022_10_15_025112_create_datos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->engine = 'InnoDB';
             $table->id();
             $table->string("RazonSocial");
-            $table->integer("Ruc");
+            $table->BigInteger("Ruc");
             $table->string("Direccion");
             $table->binary("Logo");
             $table->timestamps();
